Rename data classes to match Picaso naming Separate printer.py into several files
