+++ src/pxcli/main.py
@@ -6,7 +6,7 @@ import sys
 from time import sleep
 
 from pxctl.notifications import Notifications
-from pxctl.printer import Connection, PrinterService
+from pxctl.printer_service import Connection, PrinterService
 from layout import JsonLayout, TableLayout
 
 cur_dir = os.path.realpath(os.path.dirname(os.path.abspath(__file__)))
