+++ src/pxctl/__init__.py
@@ -1,3 +1,4 @@
 from . import notifications
-from . import printer
- 
+from . import printer_service
+from . import structs
+from . import enums
